Crate krml_reals

Reals module

Overview

The Reals library implements a new data type for safe fixed-point arithmetic. It works by creating a struct containing only an option of an i64 and then doing operator overloading for the most common arithmetic operations (addition, subtraction, multiplication, division). It also implements some convenience functions like negation, absolute value and creating from an i64. It also allows comparisons by deriving the Eq and Ord traits.

Fixed-point arithmetic

Fixed point arithmetic works similarly to normal integer arithmetic but every number is scaled by the same number, which we call the scale factor. This library allows to change the scale factor by simply changing the constant SF. By default it is set to 1 billion, which gives reals with 9 decimal points. Almost all operations works equally to integer arithmetic, except for multiplication and division. In multiplication and division the result of the operation needs to be rescaled and rounded. The range allowed by a real (for the default SF) is [-9223372036.854775808, 9223372036.854775807], which is simply the range of an i64 but rescaled.

Safe arithmetic

This library also implements safe math. All reals are an option of an i64, so a real can have the value 'None'. And all operations check for over/underflow and will return a 'None' as a result when that happens. A quirk is that, when comparing two reals, 'None' is considered smaller than any number.

Structs

Real
The struct that implements the real data type. It is a tuple containing a single Option of an i64.